// quit-smoking-app/src/domain/enums.ts
export type SmokingType =
  | "stress"
  | "social"
  | "bored"
  | "automatic"
  | "triedBefore";

export type BenefitType =
  | "calm"
  | "break"
  | "focus"
  | "comfort"
  | "identity"
  | "control";

export type TriggerType =
  | "stress"
  | "coffee"
  | "after_meal"
  | "driving"
  | "social"
  | "bored"
  | "withdrawal";